refactor: 파일 삭제, 이름 변경
TestHome.tsx -> Home.tsx
