+++ src/pages/home/Home.tsx
@@ -1,132 +1,63 @@
-import { useQuery } from '@tanstack/react-query';
-import { useEffect, useState } from 'react';
-import speciesState, { SpeciesType } from 'recoil/speciesState';
-import regionState, { RegionType } from 'recoil/regionState';
-import { useRecoilState } from 'recoil';
-import VHome, { HomeProps } from './VHome';
+import { useInfiniteQuery } from '@tanstack/react-query';
+import { useState } from 'react';
+import 'swiper/css';
+import 'swiper/css/pagination';
+import HomeVideoSlider, { HomeVideoSliderProps } from './HomeVideoSlider';
 
 const Home = () => {
-  const [currentPage, setCurrentPage] = useState(1);
-  const [shortForm, setShortForm] = useState<HomeProps | null>(null);
-  const [region, setRegion] = useRecoilState<RegionType>(regionState);
-  const [species, setSpecies] = useRecoilState<SpeciesType>(speciesState);
-
-  const fetchShortForm = async (page: number) => {
-    let type = '';
-    if (species === '강아지') {
-      type = 'DOG';
-    } else if (species === '고양이') {
-      type = 'CAT';
-    } else if (species === '기타') {
-      type = 'ETC';
-    } else {
-      type = '';
-    }
-    let area = '';
-    if (region === '전국') {
-      area = '';
-    } else {
-      area = region;
-    }
-    const apiUrl =
-      type !== '' || area !== ''
-        ? `${process.env.REACT_APP_URI}/short-forms?type=${type}&area=${area}&page=${page}&size=10`
-        : `${process.env.REACT_APP_URI}/short-forms/home?page=${page}`;
-    const response = await fetch(apiUrl, {
-      method: 'GET',
-      headers: {
-        'Content-Type': 'application/json',
+  const [muted, setMuted] = useState(true);
+  const [opacity, setOpacity] = useState(0);
+  const { data, fetchNextPage } = useInfiniteQuery(
+    ['home', 1],
+    ({ pageParam = 1 }) => {
+      const apiUrl = `${process.env.REACT_APP_URI}/short-forms/home?page=${pageParam}&size=5`;
+      return fetch(apiUrl, {
+        method: 'GET',
+        headers: {
+          'Content-Type': 'application/json',
+        },
+      }).then((res) => res.json());
+    },
+    {
+      getNextPageParam: (lastPage) => {
+        return lastPage.response.hasNext ? lastPage.response.nextPage : false;
       },
-    });
-    console.log(apiUrl);
-    if (!response.ok) {
-      throw new Error(`HTTP error! Status: ${response.status}`);
-    }
+      suspense: true,
+    },
+  );
 
-    const json = await response.json();
-    return json.response;
+  const homeVideoSliderProps: HomeVideoSliderProps = {
+    data,
+    muted,
+    setMuted,
+    setOpacity,
+    fetchNextPage,
   };
 
-  const { data, isLoading, isError } = useQuery({
-    queryKey: ['home', currentPage, species, region],
-    queryFn: () => fetchShortForm(currentPage),
-  });
-
-  useEffect(() => {
-    if (!isLoading && !isError && data) {
-      const pageData = {
-        hasNext: data.hasNext,
-      };
-
-      const shortFormData = data.shortForms;
-      const handleReachEnd = () => {
-        if (pageData.hasNext) {
-          setCurrentPage(currentPage + 1);
-          console.log('마지막');
-        } else {
-          setCurrentPage(1);
-        }
-      };
-
-      const homeProps: HomeProps = {
-        pageProps: pageData,
-        shortFormProps: shortFormData,
-        handleReachEnd,
-      };
-
-      setShortForm(homeProps);
-    }
-  }, [data, isLoading, isError, species, region]);
-
-  const handleRemoveFilter = (string: string) => {
-    if (string === species) {
-      setSpecies('전체');
-      setCurrentPage(1);
-    } else setRegion('전국');
-    setCurrentPage(1);
-    console.log('수정');
-  };
-
-  if (isLoading) {
-    return <div className="justify-center">로딩중</div>;
-  }
-
-  if (isError) {
-    return <div>Error: {isError}</div>;
-  }
-
-  if (shortForm) {
-    return species !== '전체' || region !== '전국' ? (
-      <div>
-        {/* 카테고리 선택 UI 요소를 추가 */}
-        <div className="flex justify-center gap-7 items-center">
-          <text className=" text-orange-400 text-lg font-semibold">
-            카테고리
-          </text>
-          <button
-            className="flex bg-orange-400 rounded-full px-5 py-2 text-white"
-            onClick={() => handleRemoveFilter(species)}
-          >
-            {species} x
-          </button>
-          <button
-            className="flex bg-orange-400 rounded-full px-5 py-2 text-white"
-            onClick={() => handleRemoveFilter(region)}
-          >
-            {region} x
-          </button>
-          <text className="text-lg font-semibold"> 친구들 </text>
-        </div>
-        <VHome {...shortForm} />
-      </div>
-    ) : (
-      <div>
-        <VHome {...shortForm} />
-      </div>
-    );
-  }
-
-  return null;
+  return (
+    <div className="overflow-hidden bg-slate-500 h-[90vh]">
+      <button
+        onClick={() => setMuted((prev) => !prev)}
+        className="absolute z-10 w-16 h-16 top-1/2 left-1/2 transform -translate-x-1/2 -translate-y-1/2"
+      >
+        {!muted && (
+          <img
+            src="/assets/images/speaker.svg"
+            alt="speaker"
+            style={{ opacity, transition: 'opacity 0.5s' }}
+          />
+        )}
+        {muted && (
+          <img
+            src="/assets/images/mute.svg"
+            alt="mute"
+            style={{ opacity, transition: 'opacity 0.5s' }}
+          />
+        )}
+      </button>
+      <HomeVideoSlider {...homeVideoSliderProps} />
+    </div>
+  );
 };
 
 export default Home;
